Move install actions to .environ
